requirements.txt
fastapi
uvicorn[standard]        # Include standard extras for better performance
gunicorn                 # Add gunicorn for Railway deployment
SQLAlchemy
python-dotenv
passlib[bcrypt]
PyJWT
psycopg2-binary
pydantic
email-validator